% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam_specs_sm.R
\docType{data}
\name{beam_specs_sm}
\alias{beam_specs_sm}
\title{Pediatric T-ALL BEAM Analysis Specs Data from COG trial AALL0434}
\format{
\subsection{\code{beam_specs_sm}}{

A data frame with 2 rows and 3 columns:
\describe{
\item{name}{Analysis name with omic and endpoint}
\item{mtx}{Name of omics matrix used in the analysis}
\item{mdl}{Regression model}
}
}
}
\source{
NA
}
\usage{
beam_specs_sm
}
\description{
The small beam.specs object used in example compute_beam_stats function.
}
\keyword{datasets}
