% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beam_stats_sm.R
\docType{data}
\name{beam_stats_sm}
\alias{beam_stats_sm}
\title{Pediatric T-ALL Clinical Data from COG trial AALL0434}
\format{
\subsection{\code{beam_stats_sm}}{

A beam.stats object, which contains the following objects
\describe{
\item{beam.stats}{A list of data.frames of association statistics for each omic-endpoint pair.}
\item{beam.specs}{A beam.specs object (data.frame with name, mtx, and mdl.)}
\item{beam.data}{The beam.data object.}
}
}
}
\source{
NA
}
\usage{
beam_stats_sm
}
\description{
The small beam.stats object used in example for compute_beam_stats function.
}
\keyword{datasets}
