% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beam_specs.R
\name{check_beam_specs}
\alias{check_beam_specs}
\title{Check that beam.specs satisfies all necessary conditions}
\usage{
check_beam_specs(beam.specs, mtx.names)
}
\arguments{
\item{beam.specs}{A data.frame with column name, mtx, and mdl}

\item{mtx.names}{A vector with the names of the data matrices (beam.data$mtx.data)}
}
\value{
A data.frame of beam.specs if all conditions satisfied, otherwise throws an error
}
\description{
Check that beam.specs satisfies all necessary conditions
}
\examples{
\donttest{
data(clinf)
data(omicdat)
data(omicann)
data(setdat)
data(specs)
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=100, seed=123)
test.specs <- check_beam_specs(specs, names(test.beam.data$mtx.data))
}

data(beam_dat)
data(beam_specs)
test_specs <- check_beam_specs(beam_specs, names(beam_dat$mtx.data))
}
