% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list_class.R
\name{check_list_class}
\alias{check_list_class}
\title{Check that each element of a list is of a required class}
\usage{
check_list_class(list.object, required.class)
}
\arguments{
\item{list.object}{A list used in BEAMR analysis}

\item{required.class}{Class for list elements, e.g. matrix}
}
\value{
Logical TRUE if list is of required class
}
\description{
Check that each element of a list is of a required class
}
\examples{
data(omicdat)
check_list_class(omicdat, "matrix")
}
