% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_Bmtx.R
\name{clean_Bmtx}
\alias{clean_Bmtx}
\title{Clean up bootstrap coefficient matrix}
\usage{
clean_Bmtx(B)
}
\arguments{
\item{B}{Matrix of bootstrap coefficients}
}
\value{
Matrix of cleaned bootstrap coefficients
}
\description{
Clean up bootstrap coefficient matrix
}
\examples{
\donttest{
data(clinf)
data(omicdat)
data(omicann)
data(setdat)
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
specs <- prep_beam_specs(beam.data=test.beam.data,
                         endpts=c("MRD29", "EFS", "OS"))
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data,
                                      beam.specs=specs)
B.mtx <- test.beam.stats$beam.stats[[1]]
B.cln <- clean_Bmtx(B.mtx)
}

data(beam_stats)
B.mtx <- beam_stats$beam.stats[[1]]
B.cln <- clean_Bmtx(B.mtx)
}
