% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_feature_pvalues.R
\name{compute_feature_pvalues}
\alias{compute_feature_pvalues}
\title{Compute feature level p-values from BEAM statistics}
\usage{
compute_feature_pvalues(beam.stats)
}
\arguments{
\item{beam.stats}{A beam.stats object, which is a list with beam.stats (the association matrices), the beam.specs, and the beam.data}
}
\value{
A list of feature level p-values, with each entry a data frame for a different omics/endpoint associaiton, with columns id, gene, beta, p, q
}
\description{
Compute feature level p-values from BEAM statistics
}
\examples{
\donttest{
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
specs <- prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "EFS"),
                         firth=TRUE)
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data,
                                      beam.specs=specs)
test.feat.pvals <- compute_feature_pvalues(beam.stats=test.beam.stats)
}
data(beam_stats)
test.feat.pvals <- compute_feature_pvalues(beam.stats=beam_stats)
}
