% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_set_data.R
\name{extend_set_data}
\alias{extend_set_data}
\title{Extend set definition data with genes on the same row separated by commas, semicolons, slashes, etc}
\usage{
extend_set_data(set.data, sep)
}
\arguments{
\item{set.data}{A data frame with set definition data.}

\item{sep}{Punctuation to split on.}
}
\value{
A data frame.
}
\description{
Extend set definition data with genes on the same row separated by commas, semicolons, slashes, etc
}
\examples{
data(setdat)
extend_set_data(setdat, sep=",")
}
