% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_id_clm.R
\name{find_id_clm}
\alias{find_id_clm}
\title{Find the column of mtch.data with the most rows containing an element of ids}
\usage{
find_id_clm(mtch.data, ids)
}
\arguments{
\item{mtch.data}{A data.frame}

\item{ids}{A vector of row ids to match}
}
\value{
A vector of column names with the most matches.
}
\description{
Find the column of mtch.data with the most rows containing an element of ids
}
\examples{
data(omicann)
data(omicdat)
lsn.data <- omicann[[1]]
mtx.rows <- rownames(omicdat[[1]])
test <- find_id_clm(lsn.data,mtx.rows)
}
