% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id_index.R
\name{get_id_index}
\alias{get_id_index}
\title{For each row of the data.frame main.data, find the index of the matching element in vector ids}
\usage{
get_id_index(mtch.data, ids, warn = TRUE)
}
\arguments{
\item{mtch.data}{A data.frame to be linked with the ids}

\item{ids}{A vector of ids to be linked in mtch.data}

\item{warn}{A logical value whether to include warnings with results}
}
\value{
A data.frame with matching id index
}
\description{
For each row of the data.frame main.data, find the index of the matching element in vector ids
}
\examples{
data(clinf)
data(omicdat)
mtx.clms <- colnames(omicdat[[1]])
id_index <- get_id_index(clinf,mtx.clms)
}
