% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicdat.R
\docType{data}
\name{omicdat}
\alias{omicdat}
\title{Pediatric T-ALL Omics Data from COG trial AALL0434}
\format{
\subsection{\code{omicdat}}{

A list with two dataframes of omic data for each subject
\describe{
\item{Lesion}{A dataframe with 20 rows and 265 columns indicating presence of lesion.}
\item{RNA}{A dataframe with 20 rows and 265 columns with expression data.}
}
}
}
\source{
\url{https://www.nature.com/articles/ng.3909}
}
\usage{
omicdat
}
\description{
A subset of genomic lesion and RNA expression data from pediatric and young
adult t-lineage acute lymphoblastic leukmia patients in the Children's
Oncology Group trial AALL0434, published in Liu et al., 2017 Nature
Genetics
}
\keyword{datasets}
