% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.beam.stats}
\alias{print.beam.stats}
\title{Print summary information about beam.stats object}
\usage{
\method{print}{beam.stats}(x, ...)
}
\arguments{
\item{x}{An object of class "beam.stats"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
Messages about the beam.data object
}
\description{
Print summary information about beam.stats object
}
\examples{
data(beam_stats)
print(beam_stats)
}
