% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\docType{data}
\name{specs}
\alias{specs}
\title{Pediatric T-ALL BEAMR Analysis Specs Data from COG trial AALL0434}
\format{
\subsection{\code{specs}}{

A data frame with 6 rows and 3 columns:
\describe{
\item{name}{Analysis name with omic and endpoint}
\item{mtx}{Name of omics matrix used in the analysis}
\item{mdl}{Regression model}
}
}
}
\source{
NA
}
\usage{
specs
}
\description{
The beam.specs object used in example beam analyses
}
\keyword{datasets}
