% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beam_specs.R
\name{check_beam_specs}
\alias{check_beam_specs}
\title{Check that beam.specs satisfies all necessary conditions}
\usage{
check_beam_specs(beam.specs, mtx.names)
}
\arguments{
\item{beam.specs}{A data.frame with column name, mtx, and mdl}

\item{mtx.names}{A vector with the names of the data matrices (beam.data$mtx.data)}
}
\value{
A data.frame of beam.specs if all conditions satisfied, otherwise throws an error
}
\description{
Check that beam.specs satisfies all necessary conditions
}
\examples{
data(beam_dat)
data(beam_specs)
test_specs <- check_beam_specs(beam_specs, names(beam_dat$mtx.data))
}
