% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_Bmtx.R
\name{clean_Bmtx}
\alias{clean_Bmtx}
\title{Clean up bootstrap coefficient matrix}
\usage{
clean_Bmtx(B)
}
\arguments{
\item{B}{Matrix of bootstrap coefficients}
}
\value{
Matrix of cleaned bootstrap coefficients
}
\description{
Clean up bootstrap coefficient matrix
}
\examples{
data(beam_stats)
B.mtx <- beam_stats$beam.stats[[1]]
B.cln <- clean_Bmtx(B.mtx)
}
