% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_beam_stats.R
\name{compute_beam_stats}
\alias{compute_beam_stats}
\title{Compute bootstrap model coefficients for BEAM}
\usage{
compute_beam_stats(beam.data, beam.specs, stdize = TRUE)
}
\arguments{
\item{beam.data}{Result of prep.beam.data}

\item{beam.specs}{A data.frame of strings with columns name, mtx, mdl (string with R model with mtx.row)}

\item{stdize}{Logical whether to standardize (center and scale) predictors or not. Default is TRUE.}
}
\value{
A beam.stats object, which is a list with beam.stats (the association matrices), the beam.specs, and the beam.data
}
\description{
Compute bootstrap model coefficients for BEAM
}
\examples{
\dontrun{
data(clinf)
data(omicdat)
data(omicann)
data(setdat)
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
specs <- prep_beam_specs(beam.data=test.beam.data,
                         endpts=c("MRD29", "EFS", "OS"),
                         firth=TRUE)
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data,
                                      beam.specs=specs, stdize=TRUE)
}
data(beam_dat)
data(beam_specs)
test.beam.stats <- compute_beam_stats(beam.data=beam_dat,
                                      beam.specs=beam_specs, stdize=TRUE)
}
