% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_feature_pvalues.R
\name{compute_feature_pvalues}
\alias{compute_feature_pvalues}
\title{Compute feature level p-values from BEAM statistics}
\usage{
compute_feature_pvalues(beam.stats)
}
\arguments{
\item{beam.stats}{A beam.stats object, which is a list with beam.stats (the association matrices), the beam.specs, and the beam.data}
}
\value{
A list of feature level p-values, with each entry a data frame for a different omics/endpoint associaiton, with columns id, gene, beta, p, q
}
\description{
Compute feature level p-values from BEAM statistics
}
\examples{
data(beam_stats)
test.feat.pvals <- compute_feature_pvalues(beam.stats=beam_stats)
}
