% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_beam_stats.R
\name{extract_beam_stats}
\alias{extract_beam_stats}
\title{Extract beam stats for a specific set}
\usage{
extract_beam_stats(beam.stats, set.id)
}
\arguments{
\item{beam.stats}{A beam.stats object, which is a list with beam.stats (the association matrices), the beam.specs, and the beam.data}

\item{set.id}{A character of a set id name (an entry in in beam.data$set.data$set.id)}
}
\value{
A matrix with with estimated associations for each endpoint and each omic feature linked to the set
}
\description{
Extract beam stats for a specific set
}
\examples{
data(beam_stats)
test.stats <- extract_beam_stats(beam_stats, set.id="ENSG00000099810")
}
