% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.beam.data.R
\name{print.beam.data}
\alias{print.beam.data}
\title{Print summary information about a beam.data object}
\usage{
\method{print}{beam.data}(x, ...)
}
\arguments{
\item{x}{An object of class "beam.data"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
Messages about the beam.data object
}
\description{
Print summary information about a beam.data object
}
\examples{
data(beam_dat)
print(beam_dat)
}
