% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.beam.stats}
\alias{print.beam.stats}
\title{Print summary information about beam.stats object}
\usage{
\method{print}{beam.stats}(x, ...)
}
\arguments{
\item{x}{An object of class "beam.stats"}

\item{...}{Other arguments passed to or from other methods}
}
\value{
Messages about the beam.data object
}
\description{
Print summary information about beam.stats object
}
\examples{
\donttest{
data(clinf)
data(omicdat)
data(omicann)
data(setdat)
test.beam.data <- prep_beam_data(main.data=clinf, mtx.data=omicdat,
                                 mtx.anns=omicann, set.data=setdat,
                                 set.anns=NULL, n.boot=10, seed=123)
specs <- prep_beam_specs(beam.data=test.beam.data, endpts=c("MRD29", "EFS", "OS"))
test.beam.stats <- compute_beam_stats(beam.data=test.beam.data, beam.specs=specs)
print(test.beam.stats)
}
data(beam_stats)
print(beam_stats)
}
